+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
+## [0.1.3] - 2023-08-26
+
 - Rename `TargetArch` to `Arch`, `TargetEnv` to `Env`, `TargetOs` to `Os` for consistency with `TargetSpec` field names.
 
   The old names are kept and can be used as aliases.
@@ -28,7 +30,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 Initial release
 
-[Unreleased]: https://github.com/taiki-e/target-spec-json/compare/v0.1.2...HEAD
+[Unreleased]: https://github.com/taiki-e/target-spec-json/compare/v0.1.3...HEAD
+[0.1.3]: https://github.com/taiki-e/target-spec-json/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/taiki-e/target-spec-json/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/taiki-e/target-spec-json/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/taiki-e/target-spec-json/releases/tag/v0.1.0